rename compounding simulator schema file
