+++ apps/ui/stratify-ui/app/(screens)/app/learn/compounding/compoundingSimulatorSchema.ts
